requirements.txt
Django==2.0.3
djangorestframework==3.4.6
